requirements.txt
scrapy==2.11.0
gspread==5.12.0
google-auth==2.23.4
gspread-dataframe==3.3.1
pandas==2.1.1
boto3==1.34.51